requirements.txt
boto3==1.20.26
botocore==1.23.26
jmespath==0.10.0
psycopg2-binary==2.9.2
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
